+++ CHANGELOG.md
@@ -20,6 +20,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 - BulletGraph function that deferred to BulletGraphs
 - labelWidth from BulletBuilderScope, use BulletGraphScope.labelWidth instead
+- io.github.koalaplot.core.xychart.* classes and functions which were deprecated in 0.5.0
 
 ## [0.5.2]